Added fixes for PHP7.2+. vQmod installer added.

+++ upload/install/model/install.php
@@ -1,61 +1,59 @@
 <?php
 class ModelInstall extends Model {
-	public function mysql($data) {
-		$connection = mysql_connect($data['db_host'], $data['db_user'], $data['db_password']);
-		
-		mysql_select_db($data['db_name'], $connection);
-		
-		mysql_query("SET NAMES 'utf8'", $connection);
-		mysql_query("SET CHARACTER SET utf8", $connection);
-		
-		$file = DIR_APPLICATION . 'opencart.sql';
-	
-		if ($sql = file($file)) {
-			$query = '';
+    public function mysql($data) {
+        $connection = mysqli_connect($data['db_host'], $data['db_user'], $data['db_password'], $data['db_name']);
 
-			foreach($sql as $line) {
-				$tsl = trim($line);
+        if (!$connection) {
+            die('Error: Could not connect to the database. ' . mysqli_connect_error());
+        }
 
-				if (($sql != '') && (substr($tsl, 0, 2) != "--") && (substr($tsl, 0, 1) != '#')) {
-					$query .= $line;
-  
-					if (preg_match('/;\s*$/', $line)) {
-						$query = str_replace("DROP TABLE IF EXISTS `oc_", "DROP TABLE IF EXISTS `" . $data['db_prefix'], $query);
-						$query = str_replace("CREATE TABLE `oc_", "CREATE TABLE `" . $data['db_prefix'], $query);
-						$query = str_replace("INSERT INTO `oc_", "INSERT INTO `" . $data['db_prefix'], $query);
-						
-						$result = mysql_query($query, $connection); 
-  
-						if (!$result) {
-							die(mysql_error());
-						}
-	
-						$query = '';
-					}
-				}
-			}
-			
-			mysql_query("SET CHARACTER SET utf8", $connection);
-	
-			mysql_query("SET @@session.sql_mode = 'MYSQL40'", $connection);
-		
-			mysql_query("DELETE FROM `" . $data['db_prefix'] . "user` WHERE user_id = '1'");
-		
-			mysql_query("INSERT INTO `" . $data['db_prefix'] . "user` SET user_id = '1', user_group_id = '1', username = '" . mysql_real_escape_string($data['username']) . "', salt = '" . mysql_real_escape_string($salt = substr(md5(uniqid(rand(), true)), 0, 9)) . "', password = '" . mysql_real_escape_string(sha1($salt . sha1($salt . sha1($data['password'])))) . "', status = '1', email = '" . mysql_real_escape_string($data['email']) . "', date_added = NOW()", $connection);
+        mysqli_set_charset($connection, 'utf8');
 
-			mysql_query("DELETE FROM `" . $data['db_prefix'] . "setting` WHERE `key` = 'config_email'", $connection);
-			mysql_query("INSERT INTO `" . $data['db_prefix'] . "setting` SET `group` = 'config', `key` = 'config_email', value = '" . mysql_real_escape_string($data['email']) . "'", $connection);
-			
-			mysql_query("DELETE FROM `" . $data['db_prefix'] . "setting` WHERE `key` = 'config_url'", $connection);
-			mysql_query("INSERT INTO `" . $data['db_prefix'] . "setting` SET `group` = 'config', `key` = 'config_url', value = '" . mysql_real_escape_string(HTTP_OPENCART) . "'", $connection);
-			
-			mysql_query("DELETE FROM `" . $data['db_prefix'] . "setting` WHERE `key` = 'config_encryption'", $connection);
-			mysql_query("INSERT INTO `" . $data['db_prefix'] . "setting` SET `group` = 'config', `key` = 'config_encryption', value = '" . mysql_real_escape_string(md5(mt_rand())) . "'", $connection);
-			
-			mysql_query("UPDATE `" . $data['db_prefix'] . "product` SET `viewed` = '0'", $connection);
-			
-			mysql_close($connection);	
-		}		
-	}	
+        $file = DIR_APPLICATION . 'opencart.sql';
+
+        if ($sql = file($file)) {
+            $query = '';
+
+            foreach($sql as $line) {
+                $tsl = trim($line);
+
+                if (($sql != '') && (substr($tsl, 0, 2) != "--") && (substr($tsl, 0, 1) != '#')) {
+                    $query .= $line;
+
+                    if (preg_match('/;\s*$/', $line)) {
+                        $query = str_replace("DROP TABLE IF EXISTS `oc_", "DROP TABLE IF EXISTS `" . $data['db_prefix'], $query);
+                        $query = str_replace("CREATE TABLE `oc_", "CREATE TABLE `" . $data['db_prefix'], $query);
+                        $query = str_replace("INSERT INTO `oc_", "INSERT INTO `" . $data['db_prefix'], $query);
+
+                        $result = mysqli_query($connection, $query);
+
+                        if (!$result) {
+                            die(mysqli_error($connection));
+                        }
+
+                        $query = '';
+                    }
+                }
+            }
+
+            mysqli_query($connection, "SET @@session.sql_mode = 'MYSQL40'");
+
+            mysqli_query($connection, "DELETE FROM `" . $data['db_prefix'] . "user` WHERE user_id = '1'");
+
+            mysqli_query($connection, "INSERT INTO `" . $data['db_prefix'] . "user` SET user_id = '1', user_group_id = '1', username = '" . mysqli_real_escape_string($connection, $data['username']) . "', salt = '" . mysqli_real_escape_string($connection, $salt = substr(md5(uniqid(rand(), true)), 0, 9)) . "', password = '" . mysqli_real_escape_string($connection, sha1($salt . sha1($salt . sha1($data['password'])))) . "', status = '1', email = '" . mysqli_real_escape_string($connection, $data['email']) . "', date_added = NOW()");
+
+            mysqli_query($connection, "DELETE FROM `" . $data['db_prefix'] . "setting` WHERE `key` = 'config_email'");
+            mysqli_query($connection, "INSERT INTO `" . $data['db_prefix'] . "setting` SET `group` = 'config', `key` = 'config_email', value = '" . mysqli_real_escape_string($connection, $data['email']) . "'");
+
+            mysqli_query($connection, "DELETE FROM `" . $data['db_prefix'] . "setting` WHERE `key` = 'config_url'");
+            mysqli_query($connection, "INSERT INTO `" . $data['db_prefix'] . "setting` SET `group` = 'config', `key` = 'config_url', value = '" . mysqli_real_escape_string($connection, HTTP_OPENCART) . "'");
+
+            mysqli_query($connection, "DELETE FROM `" . $data['db_prefix'] . "setting` WHERE `key` = 'config_encryption'");
+            mysqli_query($connection, "INSERT INTO `" . $data['db_prefix'] . "setting` SET `group` = 'config', `key` = 'config_encryption', value = '" . mysqli_real_escape_string($connection, md5(mt_rand())) . "'");
+
+            mysqli_query($connection, "UPDATE `" . $data['db_prefix'] . "product` SET `viewed` = '0'");
+
+            mysqli_close($connection);
+        }
+    }
 }
-?>
